Changed exceptions + ApiResponses for swagger

+++ src/main/java/hexlet/code/service/implementations/LabelServiceImpl.java
@@ -3,10 +3,13 @@ package hexlet.code.service.implementations;
 import hexlet.code.dto.LabelDto;
 import hexlet.code.model.Label;
 import hexlet.code.repository.LabelRepository;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.service.interfaces.LabelService;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -16,10 +19,12 @@ import java.util.List;
 public class LabelServiceImpl implements LabelService {
 
     private LabelRepository labelRepository;
+    private final TaskRepository taskRepository;
 
     @Override
     public Label readById(Long id) {
-        return labelRepository.findById(id).orElseThrow();
+        return labelRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Zero labels with such id"));
     }
 
     @Override
@@ -36,15 +41,20 @@ public class LabelServiceImpl implements LabelService {
 
     @Override
     public Label update(Long id, LabelDto dto) {
-        Label label = labelRepository.findById(id).orElseThrow();
+        Label label = labelRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Zero labels with such id"));
         label.setName(dto.getName());
         return labelRepository.save(label);
     }
 
     @Override
     public void delete(Long id) {
-        Label label = labelRepository.findById(id).orElseThrow();
-
+        Label label = labelRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,
+                        "Cannot delete non-existing status"));
+        if (taskRepository.existsByLabelsId(id)) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "This label is associated with a task");
+        }
         labelRepository.delete(label);
     }
 }
