+++ src/main/java/hexlet/code/service/implementations/UserServiceImpl.java
@@ -2,13 +2,16 @@ package hexlet.code.service.implementations;
 
 import hexlet.code.dto.UserDto;
 import hexlet.code.model.User;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.repository.UserRepository;
 import hexlet.code.service.interfaces.UserService;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -19,9 +22,11 @@ public class UserServiceImpl implements UserService {
 
     private UserRepository userRepository;
     private PasswordEncoder passwordEncoder;
+    private final TaskRepository taskRepository;
 
     public User readById(Long id) {
-        return userRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("No user with such id"));
+        return userRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "No user with such id"));
     }
 
     public List<User> readAll() {
@@ -30,8 +35,12 @@ public class UserServiceImpl implements UserService {
 
     public void delete(Long id) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Cannot delete non-existing user"));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,
+                        "Cannot delete non-existing user"));
 
+        if (taskRepository.existsByAuthorIdOrExecutorId(id, id)) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "This user is associated with a task");
+        }
         userRepository.delete(user);
     }
 
@@ -53,7 +62,7 @@ public class UserServiceImpl implements UserService {
 
     public User update(Long id, UserDto dto) {
         if (userRepository.findById(id).isEmpty()) {
-            throw new IllegalArgumentException("Cannot update non-existing user");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Cannot update non-existing user");
         }
         User user = User.builder()
                 .firstName(dto.getFirstName())
